SWITCHYARD-2903 Camel pooled threads never die when defined in SwitchYard
RouteBuilder creates DefaultCamelContext in configure() if it's not set, but is never stopped. If the thread pool is created in the route, it belongs to that abondoned DefaultCamelContext, causes that thread pool to survive undeploy. Fixed to pass SY global CamelContext to the RouteBuilder so that the thread pool would shutdown in the SY undeploy sequence.